Made it so that if not formatted,  shell cmds will not run

+++ TSOS-2019-master/distrib/os/shell.js
@@ -74,10 +74,8 @@ var TSOS;
             this.commandList[this.commandList.length] = sc;
             sc = new TSOS.ShellCommand(this.shellRename, "rename", "<string> - Renames a file on the hard disk");
             this.commandList[this.commandList.length] = sc;
-            // sc = new ShellCommand(this.shellList,
-            //     "ls",
-            //     "Lists all files on the disk");
-            // this.commandList[this.commandList.length] = sc;
+            sc = new TSOS.ShellCommand(this.shellList, "ls", "Lists all files on the disk");
+            this.commandList[this.commandList.length] = sc;
             // ps  - list the running processes and their IDs
             // kill <id> - kills the specified process id.
             // Display the initial prompt.
@@ -365,26 +363,66 @@ var TSOS;
         shellFormat() {
             _DSDD.initializeBlocks();
             _Console.putText("Formatted Disk Drive");
+            Formatted = true;
         }
         // Creates a file on hard disk
         shellCreate(args) {
-            _DSDD.createFile(args[0]);
+            if (Formatted) {
+                _DSDD.createFile(args[0]);
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
         }
         // Writes to a file on the hard disk
         shellWrite(args) {
-            _DSDD.writeIntoBlock(args[0], args[1]);
+            if (Formatted) {
+                _DSDD.writeIntoBlock(args[0], args[1]);
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
         }
         // Reads a file on the hard disk
         shellRead(args) {
-            _DSDD.readFromBlockUsingFilename(args[0]);
+            if (Formatted) {
+                _DSDD.readFromBlockUsingFilename(args[0]);
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
         }
         //Deletes a file on hard disk
         shellDelete(args) {
-            _DSDD.deleteFile(args[0]);
+            if (Formatted) {
+                _DSDD.deleteFile(args[0]);
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
         }
         // Renaming a file on the hard disk
         shellRename(args) {
-            _DSDD.renameFile(args[0], args[1]);
+            if (Formatted) {
+                _DSDD.renameFile(args[0], args[1]);
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
+        }
+        shellList() {
+            if (Formatted) {
+                _Console.putText("Current Files: ");
+                _Console.advanceLine();
+                let list = _DSDD.listFiles();
+                for (let i = 0; i < list.length; i++) {
+                    _Console.putText(list[i]);
+                    _Console.advanceLine();
+                }
+            }
+            else {
+                _Console.putText("Please format the drive first!");
+            }
         }
     }
     TSOS.Shell = Shell;
